requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
exceptiongroup==1.2.0
Flask==3.0.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
openai==1.3.5
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
tiktoken==0.5.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
